requirements.txt
pypdf
langchain
langchain-community
langchain-openai
openai
jinja2
python-dotenv
python-multipart
openai
faiss-cpu
uvicorn